Refactor/1: change directory, create customHooks
